cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx peterbell10 ipiszy yf225 chenyang78 kadeng muchulee8 ColinPeppler amjames desertfire chauhang

[ghstack-poisoned]

+++ torch/csrc/distributed/c10d/TCPStore.hpp
@@ -30,10 +30,10 @@ class Counter {
     return count_;
   }
   double variance() const noexcept {
-    return m2_ / count_;
+    return m2_ / static_cast<double>(count_);
   }
   double sample_variance() const noexcept {
-    return m2_ / (count_ - 1);
+    return m2_ / static_cast<double>(count_ - 1);
   }
 
  private:
